Skip objects that exist

+++ immune/objects/v5.0/reg_comboRNA/Recluster_batches_merge_RNA_v5.0.R
@@ -130,32 +130,33 @@ print(table(int$Batches))
 rna.split <- SplitObject(int, split.by = 'Batches')
 
 walk(rna.split, function(obj) {
-  obj <- runAllNormalization(obj, dims = 40)
+  if (!file.exists(glue::glue('PanImmune_{batch}_merged_RNA_{add_filename}.rds'))) {
+    obj <- runAllNormalization(obj, dims = 40)
+    batch <- obj$Batches %>% unique
+    print(batch)
+    saveRDS(obj, glue::glue('PanImmune_{batch}_merged_RNA_{add_filename}.rds'))
+    p2 <- DimPlot(obj, group.by = cell_type_column, label = TRUE)
+    ggsave(glue::glue('Dimplot_{batch}_{add_filename}_{cell_type_column}.pdf'), plot = p2, width = 12, height = 6)
+    
+    p2 <- DimPlot(obj, group.by = "Cancer", cols = colors$Cancer, label = TRUE)
+    ggsave(glue::glue('Dimplot_{batch}_{add_filename}_Cancer.pdf'), plot = p2, width = 5.5, height = 4.5)
+    
+    p2 <- DimPlot(obj, group.by = "Sample_type", cols = 'Paired', label = TRUE)
+    ggsave(glue::glue('Dimplot_{batch}_{add_filename}_Sample_type.pdf'), plot = p2, width = 5.5, height = 4.5)
+    
+    p2 <- DimPlot(obj, group.by = "seurat_clusters", label = TRUE)
+    ggsave(glue::glue('Dimplot_{batch}_{add_filename}_seurat_clusters.pdf'), plot = p2, width = 6, height = 4.5)
+    
+    p2 <- DimPlot(obj, group.by = "data.type", cols = 'Paired', label = TRUE)
+    ggsave(glue::glue('Dimplot_{batch}_{add_filename}_data.type.pdf'), plot = p2, width = 5.5, height = 4.5)
+    
+    p2 <- DimPlot(obj, group.by = "Piece_ID", label = TRUE)
+    ggsave(glue::glue('Dimplot_{batch}_{add_filename}_Piece_ID.pdf'), plot = p2, width = 12, height = 7)
+    
+    fwrite(cbind(Embeddings(obj, reduction='umap'), obj@meta.data), glue::glue('Panimmune_{batch}_merged_{add_filename}_metadata.tsv'),
+           sep='\t', row.names = T)
+  }
   
-  batch <- obj$Batches %>% unique
-  
-  saveRDS(obj, glue::glue('PanImmune_{batch}_merged_RNA_{add_filename}.rds'))
-  p2 <- DimPlot(obj, group.by = cell_type_column, label = TRUE)
-  ggsave(glue::glue('Dimplot_{batch}_{add_filename}_{cell_type_column}.pdf'), plot = p2, width = 12, height = 6)
-  
-  p2 <- DimPlot(obj, group.by = "Cancer", cols = colors$Cancer, label = TRUE)
-  ggsave(glue::glue('Dimplot_{batch}_{add_filename}_Cancer.pdf'), plot = p2, width = 5.5, height = 4.5)
-  
-  p2 <- DimPlot(obj, group.by = "Sample_type", cols = 'Paired', label = TRUE)
-  ggsave(glue::glue('Dimplot_{batch}_{add_filename}_Sample_type.pdf'), plot = p2, width = 5.5, height = 4.5)
-  
-  p2 <- DimPlot(obj, group.by = "seurat_clusters", label = TRUE)
-  ggsave(glue::glue('Dimplot_{batch}_{add_filename}_seurat_clusters.pdf'), plot = p2, width = 6, height = 4.5)
-  
-  p2 <- DimPlot(obj, group.by = "data.type", cols = 'Paired', label = TRUE)
-  ggsave(glue::glue('Dimplot_{batch}_{add_filename}_data.type.pdf'), plot = p2, width = 5.5, height = 4.5)
-  
-  p2 <- DimPlot(obj, group.by = "Piece_ID", label = TRUE)
-  ggsave(glue::glue('Dimplot_{batch}_{add_filename}_Piece_ID.pdf'), plot = p2, width = 12, height = 7)
-  
-
-  fwrite(cbind(Embeddings(obj, reduction='umap'), obj@meta.data), glue::glue('Panimmune_{batch}_merged_{add_filename}_metadata.tsv'),
-         sep='\t', row.names = T)
   
 })
